source-postgres: Warn about potential restart_lsn issues
Introduces a hook for "periodic sanity checks" and uses that to
add a "is the slot restart_lsn lagging >4GB behind the confirmed
flush LSN?" check. This has no impact beyond logging that warning,
and only exists so that we can see something in the task logs
when this happens.
