+++ source-oracle/diagnostics.go
@@ -1,0 +1,68 @@
+package main
+
+import (
+	"context"
+	"reflect"
+
+	log "github.com/sirupsen/logrus"
+)
+
+func (db *oracleDatabase) ReplicationDiagnostics(ctx context.Context) error {
+	var query = func(q string) {
+		log.WithField("query", q).Info("replication diagnostics")
+		var rows, err = db.conn.QueryContext(ctx, q)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"query": q,
+				"err":   err,
+			}).Error("unable to execute diagnostics query")
+			return
+		}
+		defer rows.Close()
+
+		var numResults int
+		cols, err := rows.Columns()
+		if err != nil {
+			log.WithFields(log.Fields{
+				"query": q,
+				"err":   err,
+			}).Error("unable to execute diagnostics query")
+		}
+		colTypes, err := rows.ColumnTypes()
+		if err != nil {
+			log.WithFields(log.Fields{
+				"query": q,
+				"err":   err,
+			}).Error("unable to execute diagnostics query")
+		}
+		for rows.Next() {
+			numResults++
+			// Scan the row values and copy into the equivalent map
+			var fields = make(map[string]any)
+			var fieldsPtr = make([]any, len(cols))
+			for idx, col := range cols {
+				fields[col] = reflect.New(colTypes[idx].ScanType()).Interface()
+				fieldsPtr[idx] = fields[col]
+			}
+			if err := rows.Scan(fieldsPtr...); err != nil {
+				log.WithFields(log.Fields{
+					"query": q,
+					"err":   err,
+				}).Error("unable to scan diagnostics query")
+			}
+
+			log.WithField("result", fields).Info("replication diagnostics")
+		}
+		if numResults == 0 {
+			log.WithField("query", q).Info("replication diagnostics: no results")
+		}
+	}
+
+	query("SELECT * FROM " + db.config.Advanced.WatermarksTable)
+	query("SELECT current_scn from V$DATABASE")
+	return nil
+}
+
+func (db *oracleDatabase) PeriodicChecks(ctx context.Context) error {
+	return nil
+}
