+++ source-oracle/replication.go
@@ -8,7 +8,6 @@ import (
 	"errors"
 	"fmt"
 	"math"
-	"reflect"
 	"slices"
 	"strconv"
 	"strings"
@@ -1086,62 +1085,6 @@ func (s *replicationStream) Close(ctx context.Context) error {
 	s.conn.Close()
 	s.cancel()
 	return <-s.errCh
-}
-
-func (db *oracleDatabase) ReplicationDiagnostics(ctx context.Context) error {
-	var query = func(q string) {
-		logrus.WithField("query", q).Info("replication diagnostics")
-		var rows, err = db.conn.QueryContext(ctx, q)
-		if err != nil {
-			logrus.WithFields(logrus.Fields{
-				"query": q,
-				"err":   err,
-			}).Error("unable to execute diagnostics query")
-			return
-		}
-		defer rows.Close()
-
-		var numResults int
-		cols, err := rows.Columns()
-		if err != nil {
-			logrus.WithFields(logrus.Fields{
-				"query": q,
-				"err":   err,
-			}).Error("unable to execute diagnostics query")
-		}
-		colTypes, err := rows.ColumnTypes()
-		if err != nil {
-			logrus.WithFields(logrus.Fields{
-				"query": q,
-				"err":   err,
-			}).Error("unable to execute diagnostics query")
-		}
-		for rows.Next() {
-			numResults++
-			// Scan the row values and copy into the equivalent map
-			var fields = make(map[string]any)
-			var fieldsPtr = make([]any, len(cols))
-			for idx, col := range cols {
-				fields[col] = reflect.New(colTypes[idx].ScanType()).Interface()
-				fieldsPtr[idx] = fields[col]
-			}
-			if err := rows.Scan(fieldsPtr...); err != nil {
-				logrus.WithFields(logrus.Fields{
-					"query": q,
-					"err":   err,
-				}).Error("unable to scan diagnostics query")
-			}
-
-			logrus.WithField("result", fields).Info("replication diagnostics")
-		}
-		if numResults == 0 {
-			logrus.WithField("query", q).Info("replication diagnostics: no results")
-		}
-	}
-
-	query("SELECT * FROM " + db.config.Advanced.WatermarksTable)
-	query("SELECT current_scn from V$DATABASE")
-	return nil
 }
 
 // marshalJSONString returns the serialized JSON representation of the provided string.
