+++ source-postgres/discovery.go
@@ -17,15 +17,6 @@ import (
 
 // DiscoverTables queries the database for information about tables available for capture.
 func (db *postgresDatabase) DiscoverTables(ctx context.Context) (map[sqlcapture.StreamID]*sqlcapture.DiscoveryInfo, error) {
-	// This is an incredibly hacky place to put this logic and it isn't going to live here
-	// in the long term, but right this moment it satisfies an important production need:
-	// we need to know what the latest server XID was at the moment that the connector
-	// started up, and we need a rough estimate of the transaction rate over time. Since
-	// this discovery code gets run at startup and also runs automatically every 5 minutes,
-	// we accomplish both of those by querying and logging out the current server XID each
-	// time discovery runs.
-	queryAndLogCurrentXID(ctx, db.conn)
-
 	// Get lists of all tables, columns and primary keys in the database
 	var tables, err = getTables(ctx, db.conn, db.config.Advanced.DiscoverSchemas, db.config.Advanced.CaptureAsPartitions)
 	if err != nil {
@@ -750,14 +741,4 @@ func getColumnDescriptions(ctx context.Context, conn *pgxpool.Pool) ([]columnDes
 		descriptions = append(descriptions, desc)
 	}
 	return descriptions, rows.Err()
-}
-
-func queryAndLogCurrentXID(ctx context.Context, conn *pgxpool.Pool) {
-	var xid uint64
-	const query = "SELECT txid_snapshot_xmin(txid_current_snapshot())"
-	if err := conn.QueryRow(ctx, query).Scan(&xid); err == nil {
-		logrus.WithField("xid", xid).Info("current transaction ID")
-	} else {
-		logrus.WithError(err).Warn("error querying current transaction ID")
-	}
 }
