+++ source-sqlserver/replication.go
@@ -1518,57 +1518,6 @@ func isReplicaDatabase(ctx context.Context, conn *sql.DB) (bool, error) {
 	return isReplica, nil
 }
 
-func (db *sqlserverDatabase) ReplicationDiagnostics(ctx context.Context) error {
-	var query = func(q string) {
-		var logEntry = log.WithField("query", q)
-		logEntry.Info("running diagnostics query")
-
-		var rows, err = db.conn.QueryContext(ctx, q)
-		if err != nil {
-			logEntry.WithField("err", err).Error("unable to execute diagnostics query")
-			return
-		}
-		defer rows.Close()
-
-		cnames, err := rows.Columns()
-		if err != nil {
-			logEntry.WithField("err", err).Error("error processing query result")
-			return
-		}
-		var vals = make([]any, len(cnames))
-		var vptrs = make([]any, len(vals))
-		for idx := range vals {
-			vptrs[idx] = &vals[idx]
-		}
-
-		var numResults int
-		for rows.Next() {
-			numResults++
-			if err := rows.Scan(vptrs...); err != nil {
-				logEntry.WithField("err", err).Error("error scanning result row")
-				continue
-			}
-			var logFields = log.Fields{}
-			for idx, name := range cnames {
-				logFields[name] = vals[idx]
-			}
-			log.WithFields(logFields).Info("got diagnostic row")
-		}
-		if numResults == 0 {
-			logEntry.Info("no results")
-		}
-	}
-
-	query("SELECT * FROM sys.dm_server_services;")
-	query("SELECT * FROM sys.dm_cdc_log_scan_sessions;")
-	query("SELECT * FROM sys.dm_cdc_errors;")
-	query("EXEC msdb.dbo.sp_help_job;")
-	query("EXEC sys.sp_cdc_help_jobs;")
-	query("SELECT * FROM msdb.dbo.cdc_jobs;")
-	query("EXEC sys.sp_cdc_help_change_data_capture;")
-	return nil
-}
-
 func unmarshalJSONString(bs json.RawMessage) (string, error) {
 	if bs == nil {
 		return "", nil
